LayoutTests/imported/w3c/web-platform-tests/IndexedDB/idbobjectstore_get3-expected.txt


PASS IDBObjectStore.get() - key is a Date
